Add logging into storage-events
