+++ packages/storage-events/src/__tests__/tmp-file.log
@@ -1,0 +1,1 @@
+[{"event":"ADDED_TODO","data":{"id":1,"title":"Implement event logging"}},{"event":"ADDED_TODO","data":{"id":2,"title":"Another event added"}}]
